user.Lookup changed user.LookupSID for Windows to avoid unnecessary internal logics.
