+++ go.mod
@@ -8,9 +8,9 @@ require (
 	github.com/VirusTotal/gyp v0.9.0
 	github.com/djherbis/times v1.5.0
 	github.com/hillu/go-yara/v4 v4.3.2
-	github.com/shirou/gopsutil/v3 v3.23.1
-	github.com/stretchr/testify v1.8.2
-	golang.org/x/sys v0.8.0
+	github.com/shirou/gopsutil/v3 v3.23.8
+	github.com/stretchr/testify v1.8.4
+	golang.org/x/sys v0.12.0
 )
 
 require (
@@ -20,10 +20,11 @@ require (
 	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
+	github.com/shoenig/go-m1cpu v0.1.6 // indirect
 	github.com/stretchr/objx v0.5.0 // indirect
-	github.com/tklauser/go-sysconf v0.3.11 // indirect
-	github.com/tklauser/numcpus v0.6.0 // indirect
-	github.com/yusufpapurcu/wmi v1.2.2 // indirect
+	github.com/tklauser/go-sysconf v0.3.12 // indirect
+	github.com/tklauser/numcpus v0.6.1 // indirect
+	github.com/yusufpapurcu/wmi v1.2.3 // indirect
 	google.golang.org/protobuf v1.27.1 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
